backend/requirements.txt
django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
yfinance
requests beautifulsoup4